// WYNKWeb/src/app/Models/CompanyMasterWebModel.ts
export class Company_Master {

  CmpID: number;
  Name = '';
  Address1: string;
  Address2: string;
Address3: string;
GSTNo: string;
ContactPerson: string;
LocationID: number;
  Phone1: string;
  Phone2: string;
  EmailID: string;
  Website: string;
  IsDeleted: boolean;
  CreatedUTC: Date;
   UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;

}